+++ 05-Callbacks-Iterators/data/positiveNegatives.data.js
@@ -1,0 +1,6 @@
+export const a = 3;
+export const b = 5;
+export const c = -3;
+export const d = -5;
+export const e = 0;
+export const testArr = [3, 5, -3, -5, 0];
