+++ 05-Callbacks-Iterators/index.js
@@ -1,46 +1,4 @@
 /**
- * #1: Positives & Negatives
- *
- * Define the function `multiplyNums`.
- * `multiplyNums` accepts two numbers and returns their multiple.
- *
- * Define the function `isOddNum`.
- * `isOddNum` accepts two numbers and a function to be called within.
- * `isOddNum` returns a string containing `a`, `b`, and the multiple of both.
- * Where a times b is zero, the string reports that the multiple is zero.
- *
- * @function multiplyNums
- * @param {Number} a
- * @param {Number} b
- * @returns {Number} the result of a times b
- *
- * @function isOddNum
- * @param {Number} a
- * @param {Number} b
- * @param {Function} multiplyFn - the function to invoke with `a` and `b`.
- * @returns {String} details whether the result of `multiplyFn` was a positive or negative number, or zero.
- *
- * @example
- * isOddNum(3, 5, multiplyNums) => "3 times 5 is a positive number."
- * isOddNum(-3, 5, multiplyNums) => "-3 times 5 is a negative number."
- * isOddNum(-3, -5, multiplyNums) => "-3 times -5 is a positive number."
- * isOddNum(-3, 0, multiplyNums) => "-3 times 0 is zero."
- * isOddNum(0, 5, multiplyNums) => "0 times 5 is zero."
- */
-
-function multiplyNums(a, b) {
-  return a * b;
-}
-
-function isOddNum(a, b, multiplyFn) {
-  if (multiplyFn(a, b) > 0) {
-    return `${a} times ${b} is a positive number.`;
-  } else if (multiplyFn(a, b) < 0) {
-    return `${a} times ${b} is a negative number.`;
-  } else return `${a} times ${b} is zero.`;
-}
-
-/**
  * Define the function `startingCapsOnly`.
  * `startingCapsOnly` accepts an array of Strings; the array is always .length >= 1
  * `startingCapsOnly` returns a new array made of any words that begin with a capital letter.
@@ -159,59 +117,6 @@ function keyValidator(testObj, targetObj) {
 import { expect } from "chai";
 
 describe("05 - Callbacks and Iterators", () => {
-  describe("#1: Positives & Negatives", () => {
-    const a = 3;
-    const b = 5;
-    const c = -3;
-    const d = -5;
-    const e = 0;
-
-    describe("returns a string", () => {
-      it("when the multiple is positive", () => {
-        expect(isOddNum(a, b, multiplyNums)).to.be.a("string");
-        expect(isOddNum(c, d, multiplyNums)).to.be.a("string");
-      });
-      it("when the multiple is negative", () => {
-        expect(isOddNum(c, b, multiplyNums)).to.be.a("string");
-        expect(isOddNum(a, d, multiplyNums)).to.be.a("string");
-      });
-      it("when the multiple is zero", () => {
-        expect(isOddNum(a, e, multiplyNums)).to.be.a("string");
-        expect(isOddNum(e, b, multiplyNums)).to.be.a("string");
-      });
-    });
-
-    describe("features the multiples in the correct order in the return string", () => {
-      it("when the multiple is positive", () => {
-        expect(isOddNum(a, b, multiplyNums)).to.include(`${a} times ${b}`);
-        expect(isOddNum(c, d, multiplyNums)).to.include(`${c} times ${d}`);
-      });
-      it("when the multiple is negative", () => {
-        expect(isOddNum(c, b, multiplyNums)).to.include(`${c} times ${b}`);
-        expect(isOddNum(a, d, multiplyNums)).to.include(`${a} times ${d}`);
-      });
-      it("when the multiple is zero", () => {
-        expect(isOddNum(a, e, multiplyNums)).to.include(`${a} times ${e}`);
-        expect(isOddNum(e, b, multiplyNums)).to.include(`${e} times ${b}`);
-      });
-    });
-
-    describe("correctly reports", () => {
-      it("when the multiple is positive", () => {
-        expect(isOddNum(a, b, multiplyNums)).to.include(`positive`);
-        expect(isOddNum(c, d, multiplyNums)).to.include(`positive`);
-      });
-      it("when the multiple is negative", () => {
-        expect(isOddNum(c, b, multiplyNums)).to.include(`negative`);
-        expect(isOddNum(a, d, multiplyNums)).to.include(`negative`);
-      });
-      it("when the multiple is zero", () => {
-        expect(isOddNum(a, e, multiplyNums)).to.include(`is zero.`);
-        expect(isOddNum(e, b, multiplyNums)).to.include(`is zero.`);
-      });
-    });
-  });
-
   describe("#2: startingCapsOnly", () => {
     const startCapWords = ["Microsoft", "Tokyo", "Charlie"];
     const noCapWords = ["party", "car", "banana"];
