+++ package.json
@@ -8,7 +8,7 @@
     "test-02": "mocha './02-Conditionals/**.js'",
     "test-03": "mocha './03-Methods-and-Functions/**.js'",
     "test-04": "mocha './04-Arrays-and-Loops/**.js'",
-    "test-05": "mocha './05-Callbacks-Iterators/**.js'",
+    "test-05": "mocha './05-Callbacks-Iterators/tests/**.js'",
     "test-06": "mocha './06-Objects/**.js'",
     "test-07": "mocha './07-Objects-2/tests/**.js'",
     "test-debug": "mocha './debug/tests/**.test.js'",
